Split up document.rs
